+++ ocean/impulse_wars/impulse_wars.c
@@ -17,7 +17,7 @@ int main(void) {
     e->rewards = fastCalloc(NUM_DRONES, sizeof(float));
     e->actions = fastCalloc(NUM_DRONES * CONTINUOUS_ACTION_SIZE, sizeof(float));
     e->masks = fastCalloc(NUM_DRONES, sizeof(uint8_t));
-    e->terminals = fastCalloc(NUM_DRONES, sizeof(uint8_t));
+    e->terminals = fastCalloc(NUM_DRONES, sizeof(float));
     e->truncations = fastCalloc(NUM_DRONES, sizeof(uint8_t));
 
     rayClient *client = createRayClient();
